POC for bundling resources via webpack.config.js
Using modules should ease introduction of JS unit tests.
