+++ .gitignore
@@ -4,7 +4,7 @@
 /local.properties
 /*.iml
 **/node_modules
-app/src/main/assets/assets/js/libs/
+/app/src/main/assets/assets/js/
 app/src/main/assets/assets/css/libs/
 app/src/main/assets/assets/css/icons.css
 app/src/main/assets/assets/fonts/
