+++ app/package.json
@@ -2,8 +2,11 @@
   "name": "pmcaFilesystemServer-app",
   "version": "1.0.0",
   "description": "",
+  "private": true,
   "scripts": {
-    "build:iconfont": "icon-font-generator src/main/assets/assets/icons/*.svg -o src/main/assets/assets/fonts/ --cssfontsurl ../fonts --csspath src/main/assets/assets/css/icons.css"
+    "build:iconfont": "icon-font-generator src/main/assets/assets/icons/*.svg -o src/main/assets/assets/fonts/ --cssfontsurl ../fonts --csspath src/main/assets/assets/css/icons.css",
+    "build": "npm run build:iconfont && webpack",
+    "watch": "webpack --watch --mode=\"development\""
   },
   "author": "Johannes Schnatterer",
   "license": "MIT",
@@ -13,9 +16,12 @@
     "file-saver": "2.0.2",
     "jquery": "3.5.1",
     "jszip": "3.5.0",
+    "popper.js": "^1.16.1",
     "vanilla-lazyload": "17.1.0"
   },
   "devDependencies": {
-    "icon-font-generator": "2.1.10"
+    "icon-font-generator": "2.1.10",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
   }
 }
