+++ app/src/main/assets/index.html
@@ -120,17 +120,7 @@
     </div>
   </div>
   
-  <script type="text/javascript" src="assets/js/libs/jquery.slim.min.js"></script>
-  <script type="text/javascript" src="assets/js/libs/bootstrap.bundle.min.js"></script>
-
-  <script type="text/javascript" src="assets/js/libs/lazyload.min.js"></script>
-  <script type="text/javascript" src="assets/js/libs/jszip.min.js"></script>
-  <script type="text/javascript" src="assets/js/libs/FileSaver.min.js"></script>
-  <script type="text/javascript" src="assets/js/libs/jszip.min.js"></script>
-  
-  <script type="text/javascript" src="assets/js/loader.js"></script>
-  <script type="text/javascript" src="assets/js/modal.js"></script>
-  <script type="text/javascript" src="assets/js/download.js"></script>
+  <script type="text/javascript" src="assets/js/main.js"></script>
 </body>
 
 </html>
