+++ app/src/main/js/index.js
@@ -1,0 +1,8 @@
+require ('bootstrap');
+
+const base = "";
+let meta = {};
+
+require('../../main/js/loader').init(base, meta);
+require('../../main/js/modal')(base);
+require('../../main/js/download').init(meta)
